02_Scraping/requirements.txt
playwright>=1.40.0
chromadb>=0.4.0
sentence-transformers>=2.2.2
tqdm
pandas
openai
gradio
uvicorn
python-dotenv
google-generativeai
lxml
